// packages/react/src/types/field.types.ts
export type FieldValue = any; // eslint-disable-line

export interface FieldValidationObject {
  rule(value: FieldValue): Promise<boolean>;
  message?: string;
  deps?: any[];
}

export interface FieldProps {
  debounce?: number;
  defaultValue?: FieldValue;
  formatValue?(value: FieldValue): FieldValue;
  name: string;
  onChange?(value: FieldValue, rawValue: FieldValue): void;
  required?: boolean;
  validations?: FieldValidationObject[];
}

export interface FieldState {
  id: string;
  name: string;
  errors: (string | undefined)[];
  value: FieldValue;
  isPristine: boolean;
}

export interface UseFieldValues {
  errorMessage?: string | undefined;
  errorMessages?: (string | undefined)[];
  id: string;
  isPristine: boolean;
  isSubmitted: boolean;
  isValid: boolean;
  setValue(value: FieldValue): void;
  value: FieldValue;
}